requirements.txt
PyYAML==6.0
RPi.GPIO==0.7.1